\name{cac40}
\docType{data}
\alias{cac40}
\alias{cac40.df}

\title{
  CAC 40 Stock Market Index (France)
}

\description{
  This \code{timeSeries} data set provides the daily closing values of the
  French CAC 40 stock index for the period 1994 to March 2004. In
  addition, the data set is also made available as a \code{data.frame}.
}

\usage{
data(cac40)
data(cac40.df)
}

\examples{
data(cac40)
head(cac40)  
}

\keyword{datasets}


