\name{danish}
\docType{data}
\alias{danish}
\alias{danish.df}

\title{
  Danish Fire Losses
}

\description{
  The \code{danish} timeSeries dataset provides the daily closing value
  for the Danish fire losses measured from January 1980 through December
  1990. In addition, the data set is also made available as a
  \code{data.frame}. 
}

\usage{
data(danish)
data(danish.df)
}

\examples{
data(danish)
head(danish)  
}

\keyword{datasets}


