\name{DJ}
\docType{data}
\alias{DJ}
\alias{DJ.df}

\title{
  Dow Jones 30 Stock Prices
} 
\description{
  The \code{DJ} timeSeries data set provides the closing values of the
  Dow Jones 30 Stocks from 1991-2000. In addition, the data set is also
  made available as a \code{data.frame}. 
}

\usage{
data(DJ)
data(DJ.df)
}

\examples{
data(DJ)
head(DJ)  
}

\keyword{datasets}


