\name{dji}
\docType{data}
\alias{dji}
\alias{dji.df}

\title{
  Dow Jones Index
}
\description{
  The \code{dji} timeSeries dataset provides the daily closing value for
  the Dow Jones index from January 1980 to March 2004. In addition, the
  data set is also made available as a \code{data.frame}. 
}

\usage{
data(dji)
data(dji.df)
}

\examples{
data(dji)
head(dji)  
}

\keyword{datasets}
