\name{ftse100}
\docType{data}
\alias{ftse100}
\alias{ftse100.df}


\title{
  FTSE 100 Stock Market Index
}

\description{
  The \code{ftse100} timeSeries dataset provides the daily closing value
  for the FTSE index from January 1980 to March 2004. In addition, the
  data set is also made available as a \code{data.frame}. 
}

\usage{
data(ftse100)
data(ftse100.df)
}

\examples{
data(ftse100)
head(ftse100)  
}

\keyword{datasets}
