\name{FXGBP.RAW}
\docType{data}
\alias{FXGBP}
\alias{FXGBP.df}

\title{
  Sterling Exchange Rates 
}

\description{
  The \code{FXGBP} timeSeries dataset provides daily exchange rates for
  major currencies (US Dollar, Japanese Yen, Euro, Swiss franc) against
  the British Pound for the period January 1987 through March 2004. In
  addition, the data set is also made available as a \code{data.frame}.  
}

\usage{
data(FXGBP)
data(FXGBP.df)
}

\examples{
data(FXGBP)
tail(FXGBP)  
}

\keyword{datasets}
