\name{hsi}
\docType{data}
\alias{hsi}
\alias{hsi.df}

\title{
  Hang Seng Stock Market Index 
}

\description{
  The \code{hsi} timeSeries dataset provides the daily closing value for
  the Hanh Seng Index from January 1994 to March 2004. In addition, the
  data set is also made available as a \code{data.frame}.
}

\usage{
data(hsi)
data(hsi.df)
}

\examples{
data(hsi)
head(hsi)  
}

\keyword{datasets}

