\name{nasdaq}
\docType{data}
\alias{nasdaq}
\alias{nasdaq.df}

\title{
  NASDAQ Stock Market Index
}

\description{
  The \code{nasdaq} timeSeries dataset provides the daily closing value
  for the NASDAQ index from January 1994 to March 2004. In addition, the
  data set is also made available as a \code{data.frame}. 
}

\usage{
data(nasdaq)
data(nasdaq.df)
}

\examples{
data(nasdaq)
head(nasdaq)  
}

\keyword{datasets}

