\name{nikkei}
\docType{data}
\alias{nikkei}
\alias{nikkei.df}

\title{
  Nikkei Stock Market Index
}

\description{
  The \code{nikkei} timeSeries dataset provides the daily closing value
  for the Nikkei index from January 1994 to March 2004. In addition, the
  data set is also made available as a \code{data.frame}. 
}

\usage{
data(nikkei)
data(nikkei.df)
}

\examples{
data(nikkei)
head(nikkei)  
}

\keyword{datasets}

