\name{smi}
\docType{data}
\alias{smi}
\alias{smi.df}

\title{
  Swiss Market Index
}

\description{
  The \code{smi} timeSeries dataset provides the daily closing value for
  the Swiss Market index from November 1990 to March 2004. In addition,
  the data set is also made available as a \code{data.frame}. 
}

\usage{
data(smi)
data(smi.df)
}

\examples{
data(smi)
head(smi)  
}

\keyword{datasets}

