\name{sp500}
\docType{data}
\alias{sp500}
\alias{sp500.df}


\title{
  Standard and Poors 500 Index
}

\description{
  The \code{sp500} timeSeries dataset provides the daily closing value
  for the S and P 500 Index from January 1980 to March 2004. In
  addition, the data set is also made available as a \code{data.frame}. 
}

\usage{
data(dji)
data(dji.df)
}

\examples{
data(sp500)
head(sp500)  
}

\keyword{datasets}
