\name{spdata.raw}
\docType{data}
\alias{spdata.raw}
\alias{spdata.raw.df}

\title{
  Standard and Poors Default Data
}

\description{
  The \code{spdata.raw} timeSeries contains default data for A, BBB, BB,
  B and C-rated companies for the years 1981 to 2000.  
}

\usage{
data(spdata.raw)
data(spdata.raw.df)
}

\source{
Standard & Poors Credit Monitor 
}

\examples{
data(spdata.raw)
head(spdata.raw)
}

\keyword{datasets}

