\name{spdata}
\docType{data}
\alias{spdata}
\alias{spdata.df}

\title{
  Standard and Poors Default Data 
}

\description{
  The \code{spdata} timeSeries dataset contains default data for A, BBB,
  BB, B and C-rated companies for the years 1981 to 2000. In addition,
  the data set is also made available as a \code{data.frame}.   
}

\usage{
data(spdata)
data(spdata.df)
}

\source{
  Standard and Poors Credit Monitor 
}

\examples{
data(spdata)
head(spdata)
}

\keyword{datasets}

