\name{xdax}
\docType{data}
\alias{xdax}
\alias{xdax.df}

\title{
  Xetra DAX German Index
}

\description{
  The \code{xdax} timeSeries dataset provides the daily closing value
  for the German Xextra DAX index from January 1994 to March 2004. In
  addition, the data set is also made available as a \code{data.frame}. 
}

\usage{
data(xdax)
data(xdax.df)
}

\examples{
data(xdax)
head(xdax)  
}

\keyword{datasets}

